Moved functions and models around to match architechture
