+++ functions/alarmLog/GetAlarmLogs.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AlarmSystem.Core.Entity.Dto;
 using core.application;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -9,21 +7,22 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 
-namespace functions.alarm
+namespace AlarmSystem.Functions.AlarmLog
 {
-    public class GetAlarmLog
+    public class GetAlarmLogs
     {
         private IAlarmLogService _alarmService;
-        public GetAlarmLog(IAlarmLogService alarmService) {
+        public GetAlarmLogs(IAlarmLogService alarmService) 
+        {
             _alarmService = alarmService;
         }
 
         [FunctionName("GetAlarmLog")]
         public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "alarmlog")] HttpRequest req,
-            ILogger log) {
-                
-                List<AlarmLog> alarmLog =  _alarmService.GetAlarmLog();
-                return new OkObjectResult(alarmLog);
+            ILogger log) 
+        {
+            List<AlarmSystem.Core.Entity.Dto.AlarmLog> alarmLogs =  _alarmService.GetAlarmLog();
+            return new OkObjectResult(alarmLogs);
         }
     }
 }
